+++ changelog.md
@@ -73,6 +73,8 @@ This version is not yet released. If you are reading this on the website, then t
 - Remove previously deprecated function strands
 ### Interpreter
 - Add the `uiua find` command, which finds Uiua code that matches the given unformatted text
+- Add checking for end-of-line signature comments
+  - These are documented in the [Documenting Code](https://uiua.org/tutorial/documentation) tutorial
 - Some optimizations
   - Pervasive function machinery has been totally rewritten
     - Observed performance improvements of up to 12x